requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
faiss-cpu==1.8.0.post1
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.8
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10
langchain-text-splitters==0.2.4
langsmith==0.1.111
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.6.0
numpy==1.26.4
opencv-python==4.10.0.84
orjson==3.10.7
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.4
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pyee==11.1.1
Pygments==2.18.0
pyparsing==3.1.4
pyppeteer==2.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
selenium==4.24.0
setuptools==72.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.33
streamlit==1.38.0
streamlit_carousel==1.1.2
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.20
watchdog==4.0.2
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.4
wheel==0.43.0
Whoosh==2.7.4
wsproto==1.2.0
yarl==1.9.7
zipp==3.20.1
